fixed pagination unique keys missing complaint
Got rid of [div wrapper[[conditional span][button]]] structure and introduced more class options in pageBtnStyle function instead, including those with pseudo elements that placed the ellipsis in the right position(s).

This eliminated the need for the "invisible" div wrappers to have unique keys, which was the cause of the problem, and the DOM is less cluttered as a result.