Pass muscle groups to Chart component
Wrap allWorkoutsMuscleGroups state destructured from useSearch in memo that depends on [allWorkoutsMuscleGroups.length] so that the Chart component does not rerender every time a page is flipped or search query is typed in search bar, but only when a workout is added or deleted.

However the Chart has to be rerendered in case the length of workouts does not change, but the muscle group data is edited in a workout, so that problem has yet to be solved.